Split speaker refinement tests into focused modules
